[#73] ParameterEqualsDefaultValueInspector: term "parameter" replaced by "argument";
