+++ src/test/java/com/kalessil/phpStorm/phpInspectionsEA/codeStyle/ArgumentEqualsDefaultValueInspectorTest.java
@@ -1,0 +1,20 @@
+package com.kalessil.phpStorm.phpInspectionsEA.codeStyle;
+
+import com.intellij.codeInsight.intention.IntentionAction;
+import com.intellij.testFramework.fixtures.CodeInsightFixtureTestCase;
+import com.kalessil.phpStorm.phpInspectionsEA.inspectors.codeSmell.ArgumentEqualsDefaultValueInspector;
+
+public final class ArgumentEqualsDefaultValueInspectorTest extends CodeInsightFixtureTestCase {
+    public void testIfFindsAllPatterns() {
+        myFixture.configureByFile("fixtures/codeStyle/argument-default-equals.php");
+        myFixture.enableInspections(ArgumentEqualsDefaultValueInspector.class);
+        myFixture.testHighlighting(true, false, true);
+
+        for (final IntentionAction fix : myFixture.getAllQuickFixes()) {
+            myFixture.launchAction(fix);
+        }
+
+        myFixture.setTestDataPath(".");
+        myFixture.checkResultByFile("fixtures/codeStyle/argument-default-equals.fixed.php");
+    }
+}
