+++ src/test/resources/fixtures/codeStyle/argument-default-equals.fixed.php
